Refactor: Move BaseNode search_by_id tests to dedicated file
